backend/requirements.txt
django-extensions
ipdb
ipython
psycopg2
python-bitcoinrpc
retry
base58
factory_boy
jsonschema